Home Page almost done and started leaderboard screen
Finishing up the Home screen just need to align everything and get rid of the white bar at the bottom of the screen. Started the leaderboards screen, I made the header and the radio buttons but I'm trying to figure out how positioning works in html so my progress has been slow so far.