Home Screen and Leaderboard Bare Min Done, Started the Game Area
Mostly finished the leaderboard Screen and home screen with a few things that could be made better but it meets my minimum requirements. I made the game area file, but I'm not sure how I should code it because I don't know how to use phaser 3.8 and how it'll appear on the page. I need to look into how to use frameworks and how to use java script to give myself some clarification.