Late Update
Since the last update, I added a few more files and filled the node module with the packages I think I need. Trying to use phaser by installing it using npm wasn't successful so I instead just downloaded the phaser.js file, but I still havent set up the actual game because I've been behind on setting up a web server. The web server I'm trying to make uses node, express and sockets. I have the server serving the html pages, but for some reason it isn't applying the css files I have working with the html. I tried getting help from the professor, but he's stumped too. The error I get is that the server doesn't want to apply the css to the html because it's a MIME type or something like that. The server is finding the file but is flat out refusing to apply the css to html.