Node Fix and Phaser Started
Finally I'm starting the development of the game itself. I was planning on making it along with the screen layouts but I needed to create a node server before that which set me back a bit because I couldn't get node use the correct directory, but after rewriting some code it finally worked. I finally found out how to make a phaser game to show on an html file. I watched a couple of tutorials, but each of them would lead me to encounter a problem I couldn't fix, however I fixed the issue I was having and now I can finally see the game show up on the html page so now I can develop it without a worry. The project's deadline is in about a month and I'm quite behind. I plan on completely focusing on this project starting april 3th because by then I won't have to worry about completing my other class's work for about 2 weeks I believe.