Changes to Scene1
Changed inputs to be based on WASD rather than a mouse drag to increase stability, added goal objects to game, added some logic in the update function for certain collisions. I'm having trouble tracking down the velocity of the puck, but to no avail. I'm going to put it off to the side and finish the other components of the game so I can start connecting my database. Not much was done because I was focusing on my other classes, but I'm in the homestretch so I need to direct my most attention to finishing this project. At the rate I've been going, I'll have the project done either this weekend or the middle of next week.