Game Almost done, Started Connecting Database
Changed the logic for moving the paddle because it would go out of bounds for some reason so instead of increasing the x and y of the paddle, the inputs control the velocity of the paddle which feel a lot more smoother than before and doesn't go out of bounds. Changed some of the logic of the puck as well, but as for the rest of the game, I'll have to revisit it later because I want to focus on connecting the database to the application. So far, the application is able to just connect to the database after reinstalling the server and manually changing the root and localhost passwords. I'm going to work on sending and receiving data using MySQL using ajax, but I'm a bit nervous because two of my other buds working on their capstone said they had no luck with ajax, but hopefully I figure out how to properly use it.