Integrated Database to Application
I was able to set up the application to connect to the database when it runs and it seems like all there's left is to work on the lines of code for sending and retrieving data, but I'm stumped on where to put the code for sending data. I know it'll go somewhere in the Scene1.js file but I'm not sure how to integrate it into the preexisting code. I'm going to get some help regarding this, but it seems like I'm close to being done.