Ajax Post Done, Ajax Get In-Progress, Cleaned Up Some Files
I spent most of today working on this project with the goal of finishing the ajax get and post operations. Sending data using Ajax works great and the database is happy after some troubleshooting, however I'm having an issue with retrieving the data using Ajax. I keep getting the error "Unexpected token '<', "<!DOCTYPE "... is not valid JSON". According to some stackoverflow discussion  boards, the problem is that Ajax is returning an html file rather than a JSON and the root of the problem for most people is the URL. Even though I know it could be the URL, I don't know what I would change the URL to.  The only thing I could think of is changing the port in the URL because the App's server runs on port 3000 while MySQL runs on port 3306. I tried to change just the port number, but that didn't work so I'm at a loss. I'm going to try to get some help at school and I'm praying that the mistake is an easy one to fix. After fixing the get with Ajax all that's left is to make everything look nice.