Praying This is the Final Update
I know I said in the last commit that it would be the last update but I noticed a huge bug in the way I did my css so I had to do some last minute editing and I structured my files and everything seems to be working fine. I'm going to test how the project runs on my laptop. If there's no further commits then I'm proud to say that I'm DONE.