Final Update this time?
In my last commit the css for the leaderboard looked fine on my pc but it didn't look good on my laptop so I was forced to revert to using pixels for declaring the height of certain elements. I really wanted to make my website more responsive, but I failed to do so. Hopefully when I send this commit to my pc everything will be fine.